Cálculos en el carrito de compras
- Se reorganizó el código para una mejor separación de responsabilidades.
- Se cambió el tipo de dato de la columna 'precio' en la base de datos de REAL a INTEGER para representar precios en pesos colombianos sin decimales.
- Se ajustaron los cálculos del subtotal, tarifa de envió y total.
- Se implementó formateo con separador de miles (ej. 1.200 en vez de 1200) para mejorar la presentación visual en los TextView.

+++ app/src/main/java/com/example/gaia/db/DbCarrito.java
@@ -62,7 +62,7 @@ public class DbCarrito extends DbHelper {
                         cursor.getInt(cursor.getColumnIndexOrThrow("id")),
                         cursor.getString(cursor.getColumnIndexOrThrow("imagen")),
                         cursor.getString(cursor.getColumnIndexOrThrow("nombre")),
-                        cursor.getDouble(cursor.getColumnIndexOrThrow("precio")),
+                        cursor.getInt(cursor.getColumnIndexOrThrow("precio")),
                         cursor.getInt(cursor.getColumnIndexOrThrow("cantidad"))
 //                        cursor.getString(cursor.getColumnIndexOrThrow("descripcion")),
 //                        cursor.getString(cursor.getColumnIndexOrThrow("ingredientes")),
