+++ app/src/main/java/com/example/gaia/models/ProductoCarrito.kt
@@ -4,7 +4,7 @@ data class ProductoCarrito(
     val id: Int,
     val imagen: String,
     val nombre: String,
-    val precio: Double,
+    val precio: Int,
     var cantidad: Int
 //    val descripcion: String,
 //    val ingredientes: String,
